+++ tests/unit/lib/date.spec.ts
@@ -4,7 +4,9 @@ import {
   getFullDayString,
   getFullDayWithTimeString,
   getDateRepresentationWithoutSameDate,
-  getDisplayDate
+  getDisplayDate,
+  compareDate,
+  compareDateString
 } from '@/lib/date'
 
 const defaultDate1 = new Date('1985-12-04T15:21:34')
@@ -83,5 +85,61 @@ describe('getDisplayDate', () => {
   it('should get date string when not same date', () => {
     const dateISO2 = '2001-06-04T08:25:34'
     expect(getDisplayDate(dateISO, dateISO2)).toEqual('06/04 08:25')
+  })
+})
+
+describe('compareDate', () => {
+  it('can compare time', () => {
+    expect(
+      compareDate(
+        new Date(2020, 1, 1, 0, 0, 0, 0),
+        new Date(2020, 2, 1, 0, 0, 0, 0)
+      )
+    ).toEqual(-1)
+  })
+  it('can compare time (inversed)', () => {
+    expect(
+      compareDate(
+        new Date(2020, 1, 1, 0, 0, 0, 0),
+        new Date(2020, 2, 1, 0, 0, 0, 0),
+        true
+      )
+    ).toEqual(1)
+  })
+  it('can compare same time', () => {
+    expect(
+      compareDate(
+        new Date(2020, 1, 1, 0, 0, 0, 0),
+        new Date(2020, 1, 1, 0, 0, 0, 0)
+      )
+    ).toEqual(0)
+  })
+})
+
+describe('compareDateString', () => {
+  it('can compare time', () => {
+    expect(
+      compareDateString(
+        '2021-02-19T17:09:24.53365Z',
+        '2021-02-20T12:16:03.559831Z'
+      )
+    ).toEqual(-1)
+  })
+  it('can compare time (inversed)', () => {
+    expect(
+      compareDateString(
+        '2021-02-19T17:09:24.53365Z',
+        '2021-02-20T12:16:03.559831Z',
+        true
+      )
+    ).toEqual(1)
+  })
+  it('can compare same time', () => {
+    expect(
+      compareDateString(
+        '2021-02-19T17:09:24.53365Z',
+        '2021-02-19T17:09:24.53365Z'
+      )
+    ).toEqual(0)
   })
 })
